preuba chi Cuadrado
Se cambian todos los tipos de datos double por float. con double los resultados dan mal. se agregan metodos de calculo de estadistico de prueba en Estadistica

+++ src/sample/model/GeneradorAleatorioJava.java
@@ -6,9 +6,9 @@ public class GeneradorAleatorioJava implements IGeneradorAleatorio
         return 0;
     }
 
-    public double GenerarAleatorio()
+    public float GenerarAleatorio()
     {
-        return Math.random();
+        return (float) Math.random();
     }
 
     public void SiguienteSemilla() {
